Schemas improvements
- Create schemas file
- Imports improvements
